requirements.txt
fastapi
uvicorn
PyMuPDF 
python-docx
chromadb
transformers
torch
openai

